Check for a required feature in libXaw or libXaw3d
Check for a required feature in libXaw or libXaw3d, and warn sternly if
the feature is not found. At least libXaw7 (or compatibles) should be
used; older versions lack features and libXaw3d based on Xaw6 is buggy.
This should really be an error, and will become so in the future.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@28584 379a1393-f5fb-40a0-bcee-ef074d9b53f7