Gtk3: display fliplist activity in the statusbar
See bug #1129. Gtk3's fliplist now shows on the statusbar what it's doing. Unfortunately the statusbar's area for messages is a little small, so not useful info is actually displayed.
Perhaps the GtkLabel can be told to use multiline, or can be added as a third row in the statusbar.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@36950 379a1393-f5fb-40a0-bcee-ef074d9b53f7