Avoid DISK FULL with 27 blocks free on .D80 / .D82.
For all drive types, BAM_TRACK_xxxx and DIR_TRACK_xxxx are the same,
except for 8050 and 8250. See <diskconstants.h>.

Previous code avoided allocating sectors on both BAM_TRACK_xxxx and
DIR_TRACK_xxxx (or vdrive->Bam_Track and vdrive->Dir_Track).

However, the 8[02]50 is perfectly able to allocate on track 38
(Bam_Track): in fact it is where it starts.

So this patch changes all occurrences of Bam_Track into Dir_Track where
it involves allocating sectors, or avoiding so. Since they are the same
for most drives, it makes no difference for them. For the 8[02]50 it
allows to allocate sectors on track 38, and so avoids "DISK FULL" errors
when there are still 27 blocks free on track 38.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@38100 379a1393-f5fb-40a0-bcee-ef074d9b53f7