Gtk3: When swapping joysticks (Alt+J), also swap the emulated device types
See feature request #410. When swapping devices of different types, for example mouse and joystick, the underlaying device types would not get swapped. This is a bug, so now we also swap the device types along with the input devices. When trying to swap a device that only works in port #1 (lightpen) the UI displays an error message and the old situation is restored.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@40511 379a1393-f5fb-40a0-bcee-ef074d9b53f7