Gtk3: Userport device combobox: add device type desc to model, split model generation into separate function
The model now contains a short string describing the device type. But I haven't found a way to add this to the combobox in a nice-looking way. A second column with device type looks a bit weird and Gtk doesn't allow "section headers" inside comboboxes. So I'll probably have to leave the combobox as-is.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@40701 379a1393-f5fb-40a0-bcee-ef074d9b53f7