Gtk3: Change the "Swap joysticks" button in the settings into a toggle button like in the menus, rename KeySetEnable toggle button to "Allow keyset joysticks"
Using a toggle button for "Swap joysticks" gives the user a visual indication the joysticks are swapped or not, like in the menus. After the release of 3.6 we should implement swapping control ports in the joystick code when a resource is set and not swap the Joy[Port]?[12]Device resources.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@41303 379a1393-f5fb-40a0-bcee-ef074d9b53f7