Add CHIPShowStatusbar resources and [-+]CHIPshowstatusbar command line options, implement handling of these in the Gtk3 UI
The status bar in the Gtk3 UI can now be shown/hidden via a toggle button in the menu and via command line options. Currently starting with the status bar hidden results in some grey bands because the status bar is shown during UI setup and hidden afterwards; can be fixed with `gtk_widget_show_no_show_all()`, but I'm tired.
Some additional tweaking might be required to resize the window when toggling the status bar visibility.
A hotkey action 'show-statusbar-toggle' has been added but no default hotkey assigned yet: x128 needs an additional action 'show-statusbar-vdc-toggle' to be able to independently toggle the status bars of the VICII and the VDC windows.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@42225 379a1393-f5fb-40a0-bcee-ef074d9b53f7