Gtk3: VSID: Disable drag'n'drop handler for the STIL GtkTextView
To be able to drop SID files onto the GtkTextView portion of the STIL pane the widget needs its own drag'n'drop events AND have 'editable' set to TRUE, which makes it possible to paste text into the widget (which we don't want).
As a bonus I've disabled the drag'n'drop handler for the playlist, since that currently triggers a nice segfault.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@42257 379a1393-f5fb-40a0-bcee-ef074d9b53f7