Gtk3: Remove ui_accel_data_t object and its methods
Since we keep a table of menu item data for the UI actions and hotkeys we can
now simply pass a pointer to an element of that table as the closure data.
We no longer need to manually allocate and destroy an object containing the data
required for the closure, the table contains all the information required for
the closure to do its job.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@42318 379a1393-f5fb-40a0-bcee-ef074d9b53f7