Fix speeddos+ cable.
This was broken in r42123 when rearranging the code for CB1 edge
detection. For some reason, limiting this to actual changes in the input
breaks this use case (which seems a bug in itself).
You can see if the parallel cable is really used: when the LOAD command
prints start and end addresses.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@42332 379a1393-f5fb-40a0-bcee-ef074d9b53f7