Gtk3: Implement UI actions
Provides a mechanism to trigger a UI action using a simple ID. This should
make it possible to trigger UI actions from a controller as well, and to add
hotkeys for UI actions that do not have a corresponding menu item (which will
be useful for VSID especially).
Since this is a work in progress, not all menu items will work, but the
basics (attach disk, warp, pause, monitor, settings dialog, reset, quit)
are working. The rest of menu items and the popup menus will be moved over
to the new system in the coming days. So if in doubt, use the commandline or
the SDL port, for the time being.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@42390 379a1393-f5fb-40a0-bcee-ef074d9b53f7