Gtk3: Add UI actions for fliplists of unit 9, 10 and 11, prepare for dual-drive
Instead of fliplist-add/remove etc assuming unit 8, drive 0, the actions now
cover units 8-11, drive 0/1. Since the fliplist code doesn't (yet) support
fliplists for dual drives only the actions for drive 0 are implemented. Action
IDs and strings are available for drive 1, but no handlers are implemented for
those.
The names of the fliplist actions in the .vhk files have changed:
'fliplist-add' has become 'fliplist-add-8:0', 'fliplist-load' has become
'fliplist-load-8:0' and so on.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@42431 379a1393-f5fb-40a0-bcee-ef074d9b53f7