Gtk3: Make handler_accelerator() return `TRUE` to avoid hotkeys ending up in the emulated machine
Although the prototype for a closure callback specifies it returning nothing,
we need to return `TRUE` because its being used in the context of an accelarator,
which needs to return gboolean. Returning `TRUE` signals the key has been
consumed and thus the key won't end up in the emulated machine.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@42438 379a1393-f5fb-40a0-bcee-ef074d9b53f7