Github Actions: Add building Debian package on commit (WIP)
Add an action to build a binary deb package of the Gtk3 UI for amd64.
Currently the package should install into /usr. The vice.pdf file is currently
not included.
The resulting package has not been tested at all, so do not use on a system
you like to keep intact =)



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@42687 379a1393-f5fb-40a0-bcee-ef074d9b53f7