Gtk3: Add workaround for Pango >1.50.11 not loading CBM font on Windows
Since 1.50.12 Pango no longer recognizes fonts loaded via GDI using
AddFontResourceEX(). In 1.50.13 a patch will be included that adds back some
of the functionality, but we have to use AddFontResourceA(), which means the
font is registered globally and if VICE crashes the system needs to be rebooted
to "uninstall" the font.
See https://gitlab.gnome.org/GNOME/pango/-/issues/720



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@42812 379a1393-f5fb-40a0-bcee-ef074d9b53f7