"Merge" branches/fabrizio/3.7-joystick-mapping into trunk
Apply diff of the branch with trunk on trunk and commit. The branch cannot be
merged with `svn merge` due to broken history in the branch, so I've created
a diff and applied the diff on trunk. Do NOT blame me if things break, I'm
just applying the changes in the branch, I didn't write them.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@42972 379a1393-f5fb-40a0-bcee-ef074d9b53f7