Gtk3: Fix hotkeys loading with `-directory` on Windows
Remove the $VICEDIR and $USERDIR special tokens in the !include directives.
The added flexibility then writing and testing .vhk files doesn't weigh up to
the complexity when sysfile.c and findpath.c get involved.
The findpath() function still has a bug with adding another dir separator when
building a path, that'll need to be investigated and fixed separately.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@43039 379a1393-f5fb-40a0-bcee-ef074d9b53f7