Gtk3: Resource-bound combo boxes: remove combo box with custom entry, remove _get() methods
The resource combo box with a custom text entry isn't used anywhere, so rather
than refactoring it to use the resource mediator it is new deleted.
Remove the _get() methods: it is not clear from looking at the method name if
it gets the resource value or the widget value, and they weren't used anywhere
in the code. They can be reimplemented as two separate methods if required.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@43074 379a1393-f5fb-40a0-bcee-ef074d9b53f7