Gtk3: Resource mediator: add generic object pointer to be used for additional state
Some widgets require additional state data to be kept, but the adding all these
properties to the mediator would make the mediator have a lot of rarely used
method and properties. And since we don't have inheritance in C we provide a
simple method to set a void* pointer and register a callback to free said
pointer on widget destruction.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@43100 379a1393-f5fb-40a0-bcee-ef074d9b53f7