Gtk3: New resource file chooser widget: open/create files/directories
Create a widget bound to a string resource to select or create files or
directories for filename and directory resources.
The widget is a GtkEntry with a clickable button inside it to open a file
chooser dialog, instead of the ugly "Browse..." button in the "resource browser"
widget. There is no option label in front of it, since that usually caused
layout issues in the old resource browser widget.
Internally the code is a lot cleaner and simpler than the old resource browser
widget it is meant to replace.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@43120 379a1393-f5fb-40a0-bcee-ef074d9b53f7