Gtk3: Settings dialog: rework MMC64 layout and code
Simplify code by removing a lot of custom widget code and static variables that
aren't used for anything (left-over from old glue logic perhaps). Try to group
the widgets together a little more consistently. Use a GtkSwitch for the jumper,
a physical button on the cart. Group the flush/save buttons in a GtkButtonBox
for homogeneuous sizes of the buttons.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@43169 379a1393-f5fb-40a0-bcee-ef074d9b53f7