Add (but disabled) change for latching from  Massimiliano Zattera (bug 1828).
Latching was and still is disabled, because apparently no known software uses it.
However I have added the suggested changes from Massimiliano Zattera.

Because the order of the code clears the interrupt flag before checking it,
I needed to add an extra variable, and to make sure it doesn't get overwritten
for the no-handshake case, a goto to skip its assignment.

Perhaps it would make more sense to reverse the order of doing the handshake
and interrupt clearing vs the reading of the port. Certainly it would make sense
to read the port first and then signal the handshake that it has been read.
But I didn't change that at this time. I would need to test if something depends
on this order.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@43238 379a1393-f5fb-40a0-bcee-ef074d9b53f7