CRTC_BEAM_RACING revealed a long-standing bug in the use of crtc.vaddr_mask.
It was used both before and after applying the chars-per-cycle factor.  As a
result, it went unnoticed that for the 8296 the vaddr_mask included one too
many bits: the one marked "Bit 4: use top half of 4K character generator".
This caused the wrong screen memory to be shown.
For other 80-column PETs there was a bit too many as well but this is not
normally noticable.
Solve it by having 2 masks. If we would ignore wrap-around, the second mask
isn't even needed.
Tested lightly with SuperPET APL, 8296 HRE, xcbm2, xcbm2 -model 710.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@43835 379a1393-f5fb-40a0-bcee-ef074d9b53f7