Don't add uiactions.c to EXTRA_DIST conditionally
It appears any files added to foo_SOURCES conditionally will always end up in
the dist tarball when doing a `make dist`, regardless of the condition in the
Makefile. I checked this with the `macOS-util.m` file: it is only added to
libarchdep_a_SOURCES when `MACOS_COMPILE` is true, yet a source tree configured
for Linux+headless will result in `macOS-util.m` added to the dist tarball.
The same is true for `SUBDIRS`, the `hotkeys` subdir is added conditionally to
the list when `!USE_HEADLESSUI`, yet the source configured for headless UI adds
the `hotkeys` subdir to the dist as well.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@44024 379a1393-f5fb-40a0-bcee-ef074d9b53f7