This should fix bug #1900.
When the machine is reset, the clock is set back too. Any outstanding CIA
alarms are suddenly quite far into the future. Most should even be turned off
(since the CIA is now reset). The idle alarm should be set to the new near future.
Do just that in ciacore_reset().



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@44076 379a1393-f5fb-40a0-bcee-ef074d9b53f7