Gtk3: Add custom spin button widgets in settings/tapeport
Add reusable spin button widget that translates between resource values and
displayed/entered values and allows for custom formatting of the resource value
via a printf()-style format string.
Currently used to display "DatasetteSpeedTuning" as a range of -50.0% to +50.0%,
"DatasetteTapeWobbleFrequency" as a range of 0.0Hz to 50.0Hz, and
"DatasetteTapeWobbleAmplitude" as a range of 0.0% to 50.0%.
Still needs some code to allocate additional horizontal spacing in the internal
GtkEntry of the GtkSpinButton to account for any affix or other non-numeric
characters in the custom formatted output.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@44331 379a1393-f5fb-40a0-bcee-ef074d9b53f7