Gtk3: Add sprintf version of custom spin button widget constructor, fix drive RPM widgets
Add a constructor for the custom spin button widget that allows using an
sprintf-style format string and arguments for that format string so we no longer
have to construct the resource name manually.
Fix drive RPM widgets: `gtk_spin_button_set_digits()` doesn't work with the
custom spin buttons, it only allocates space in the widget for the digits but
doesn't have any bearing on the number of digits displayed since we use a custom
format string (that string determines the digits). Also fix the format strings
for the spin buttons: in "%M.Nf" M doesn't mean the number of digits before the
".N", it includes it, so for 3 digits before the period and 2 after, we have to
use "%6.2f" instead of the more logical looking "%3.2f".



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@44349 379a1393-f5fb-40a0-bcee-ef074d9b53f7