Add script for github actions to compare generated dist tarballs
Add a script that unzips dist tarballs generated by the 'check-make-dist' job
and compares the file listings of each dist tarball with the others. The script
runs `diff` on the file listings of the tarballs so we can visually inspect the
result of the script, should it report differences. Exit status 0 means all
file listings are equal, 1 means some differ, and 2 means an error occurred
during execution of the script.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@44408 379a1393-f5fb-40a0-bcee-ef074d9b53f7