Gtk3: Fix saving  monitor position and size and restoring on boot
For some reason Gdk decides to restore the settings from vicerc (if present)
or the centered coordinates when the monitor window is hidden, resulting in a
call to `on_window_configure_event()` with those coordinates, wiping any
coordinates we updated the resources with when moving/resizing the window.
So we check if the window is hidden in the event handler and bail out early if
so.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@44451 379a1393-f5fb-40a0-bcee-ef074d9b53f7