Gtk3: check monitor xpos/ypos for `INT_MIN` as non-set values
On Windows and perhaps some other systems, xpos/ypos of windows can be negative,
so we should check for the special `INT_MIN` value indicating the resources are
not set instead of <0.
Also restore the gtk_window_present() call I accidentally removed while debugging
that is required for the monitor window showing in fullscreen on Windows.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@44452 379a1393-f5fb-40a0-bcee-ef074d9b53f7