Gtk3: Use eye icon for WiC64 security token visibility toggling
Instead of a separate check button use an icon inside the GtkEntry that allows
toggling the visibility of the security token. Unfortunately the right corner
of the entry is already used for the Caps Lock indicator, so we have to use the
left corner of the entry for this icon (Gtk4 has a GtkPasswordEntry that does
the correct thing, Gtk3 doesn't).


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@44574 379a1393-f5fb-40a0-bcee-ef074d9b53f7