Add an entry to the Sound menu for the CB2 lowpass filter.
For PETs only. Because there is no nice way to have some sort of conditional
menu item in a common menu, I add the item by copying the whole menu and insterting
the entry, on the PET only. Since the affected menu is attached directly to the
main menu, this doesn't need patching a whole menu tree.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@44643 379a1393-f5fb-40a0-bcee-ef074d9b53f7