ui_display_statustext(): make `fadeout` a proper `bool` instead of `int`
Somewhat against expectations, the `fadeout` argument is a boolean argument,
enabling a fadeout of 5 seconds or leaving the message indefinitely. I vaguely
recall the `fadeout` argument being an actual integer specifying the fadeout in
seconds, but a lot of code now used either 0 or 1. So to avoid confusion the
`fadeout` argument is now of type `bool`.
Also remove the duplicate prototype for the function from `uistatusbar.h`, the
prototype lives in `uiapi.h` since it's UI-agnostic.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@44929 379a1393-f5fb-40a0-bcee-ef074d9b53f7