Merge branch compyx/evdev-joystick into trunk
Add experimental Linux evdev joystick driver. Driver can be enabled by passing
`--with-evdev` to configure, which requires libevdev and its headers to be
installed. When the evdev driver is enabled the old Linux joystick driver will
be disabled since we have no mechanism in place to select a joystick driver.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@44980 379a1393-f5fb-40a0-bcee-ef074d9b53f7