Windows: copy dependencies of the GdkPixbuf SVG loader into bindist
Since the SVG loader is dynamically loaded by GdkPixbuf, the `ntldd` call on
an emulator executable won't pick it up, we need explicitly get the deps of
the SVG loader and copy them into the bindist as well.
Should fix bug #2081, please test once the bindist is built on Github.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@45390 379a1393-f5fb-40a0-bcee-ef074d9b53f7