XDG desktop files: handle make's DESTDIR to allow for staged installs
Update the calls to `desktop-file-install` to pass $DESTDIR as its `--dir`
argument so installing into a staging directory as a non-root user will place
the .desktop files into their proper location. The installation of icon files
with `xdg-icon-resource` into a staging directory will require some equivalent
logic to be applied (working on that).
Also add dashes in front of the calls to `update-desktop-database` to have
`make` ignore the failure of said program when not running as root.
See bug #2127.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@45506 379a1393-f5fb-40a0-bcee-ef074d9b53f7