XDG desktop files: use `$(DESTDIR)${prefix}` for installation
As suggested by Rhailto, use `--dir-=$(DESTDIR)${prefix}/share/applications` as
argument to `desktop-file-install` to support "normal" installation of .desktop
files through `make install` and allow for staged installs. Update the rules
to remove the .desktop files on `make uninstall` in the same vein.

With the default prefix of `/usr/local` that VICE's configure uses, the .desktop
files end up in `/usr/local/share/applications`, a directory that didn't exist
on my Debian system with Gnome, but which is created by `desktop-file-install`
and the .desktop files are processed by `update-desktop-database` and VICE's
application icons show up in the Gnome application overview and correctly launch
the emulators.

Note the use of `${prefix}` rather than `$(PREFIX)`, the later was empty on my
`make install`, resulting in .desktop files in `/share/applications` in the
root file system =)



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@45507 379a1393-f5fb-40a0-bcee-ef074d9b53f7