Re-fix access to empty memory space.
Now that I finally get why this almost always results in the high byte of the address, make sure that access to $9xxx, $Axxx and $Bxxx goes via read_unused() when no ROM is in the socket; this should take care of the exceptions. Previously there would be a ROM pre-filled with the empty value.

Also: a memory map reset would turn off watchpoints. Keep the watchpoints active in that case.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@45633 379a1393-f5fb-40a0-bcee-ef074d9b53f7