Merge branch compyx/joymap-002 intro trunk
Implement opening and closing host joystick devices via resources and manually.
Host devices will now only be opened when a "JoyDevice{N}" resource is set, or
when joystick_device_open() is called manually. When a "JoyDevice{N}" resource
is changed the old host device (if any) is closed before opening the new one.
This prepares the code for polling via the UI (for mapping and calibration).


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@45654 379a1393-f5fb-40a0-bcee-ef074d9b53f7