On the 8296 we do not (invert the screen by clearing MA12).
The CRTC code (crtc.c) which tests vaddr_revswitch (as it is called there) has no
explicit value for "we do not have the reverse feature at all", but it is clever
enough to use an uninverted screen if the toggle bit is part of the "normal"
screen addressing bits. So we just use the video address mask (vaddr_mask) bits
as the vaddr_revswitch to disable inverting.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@45677 379a1393-f5fb-40a0-bcee-ef074d9b53f7