c1541: fix segfault when not attaching disk image.
When a disk image is attached, vdrive_device_setup() is called to set up the
vdrive struct. This includes memory for the disk error string buffer.

If you do a "c1541 -write foo bar" command (so without attaching a disk image),
c1541 wants to write a DRIVE NOT READY to the error string buffer. This
segfaults because it has not been allocated.

Fix: Call vdrive_device_setup() right after a vdrive is allocated, so it is
always good to use. Also, when freeing the buffers in vdrive_device_shutdown(),
set them to NULL to make sure they are not re-used accidentally.
When vdrive_device_setup() is called again for attaching a disk image, be
paranoid and first call vdrive_device_shutdown() on it.
When detaching an image, call vdrive_device_shutdown() even if no image was
attached. This prevents memory leaks.

Also: replace some 8's by DRIVE_UNIT_MIN.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@45691 379a1393-f5fb-40a0-bcee-ef074d9b53f7