Fix bug 2175: CRTC 80x42 text mode does not display properly.
The test program set the CRTC to 42 lines of 7 scan lines each, making
294 scan lines total. This needs a PET 8296 to have sufficient screen
memory for all these character positions. This didn't fit in the standard
window size (which could fit up to 256 scanlines). As a result, various
unwritten assumptions in the code where the CRTC interacted with raster/*.c
code must have been broken, and the display was not just truncated but garbled.

We fix this by adding some code to resize the window larger when needed (and
also smaller if the required size is much smaller than the current size).
Various border sizes seemed to be applied inconsistently; this was changed.

Since the code never draws a top and bottom border, the allocation of render
bitmaps for the GTK3 GUI was changed to clear them before use. This normally
only happens once or twice at startup, and again when resizing the window
larger.  Therefore this has no performance impact.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@45801 379a1393-f5fb-40a0-bcee-ef074d9b53f7